#117 - auth, stat 도메인 기준으로 패키지 구조 변경
