--- server/module-api/src/main/java/project/BaekjoonStatus/api/argumentresolver/Auth.java
@@ -1,11 +1,0 @@
-package project.BaekjoonStatus.api.argumentresolver;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.PARAMETER)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Auth {
-}
