--- server/module-api/src/main/java/project/BaekjoonStatus/api/cache/CacheCallback.java
@@ -1,5 +1,0 @@
-package project.BaekjoonStatus.api.cache;
-
-public interface CacheCallback<T> {
-    T execute();
-}
