--- server/module-api/src/main/java/project/BaekjoonStatus/api/cache/CacheNode.java
@@ -1,16 +1,0 @@
-package project.BaekjoonStatus.api.cache;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-@Data
-public class CacheNode<T> {
-    LocalDateTime expireAt;
-    T data;
-
-    public CacheNode(LocalDateTime expireAt, T data) {
-        this.expireAt = expireAt;
-        this.data = data;
-    }
-}
