+++ server/module-api/src/main/java/project/BaekjoonStatus/api/common/argumentresolver/Auth.java
@@ -1,0 +1,11 @@
+package project.BaekjoonStatus.api.common.argumentresolver;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Auth {
+}
