+++ server/module-api/src/main/java/project/BaekjoonStatus/api/common/config/WebConfig.java
@@ -1,0 +1,38 @@
+package project.BaekjoonStatus.api.common.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import project.BaekjoonStatus.api.common.argumentresolver.AuthArgumentResolver;
+import project.BaekjoonStatus.api.common.interceptor.AuthInterceptor;
+
+import java.util.List;
+
+@Configuration
+@RequiredArgsConstructor
+public class WebConfig implements WebMvcConfigurer {
+    @Value("${token.secret}")
+    private String tokenSecret;
+
+    @Override
+    public void addArgumentResolvers(List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(new AuthArgumentResolver(tokenSecret));
+    }
+
+    @Override
+    public void addInterceptors(InterceptorRegistry registry) {
+        registry.addInterceptor(new AuthInterceptor(tokenSecret))
+                .order(1)
+                .addPathPatterns("/stat/**", "/auth/me")
+                .excludePathPatterns("/", "/auth/baekjoon", "/auth/signup", "/auth/login", "/error");
+    }
+
+    @Override
+    public void addCorsMappings(CorsRegistry registry) {
+        registry.addMapping("/**");
+    }
+}
