--- server/module-api/src/main/java/project/BaekjoonStatus/api/config/AsyncConfig.java
@@ -1,26 +1,0 @@
-package project.BaekjoonStatus.api.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.core.task.TaskExecutor;
-import org.springframework.scheduling.annotation.AsyncConfigurer;
-import org.springframework.scheduling.annotation.EnableAsync;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-
-import java.util.concurrent.Executor;
-
-@Configuration
-@EnableAsync
-public class AsyncConfig implements AsyncConfigurer {
-    @Override
-    public TaskExecutor getAsyncExecutor() {
-        ThreadPoolTaskExecutor taskExecutor = new ThreadPoolTaskExecutor();
-
-        taskExecutor.setCorePoolSize(1);
-        taskExecutor.setMaxPoolSize(1);
-        taskExecutor.setQueueCapacity(1000);
-        taskExecutor.setThreadNamePrefix("MyExecutor");
-        taskExecutor.initialize();
-
-        return taskExecutor;
-    }
-}
