+++ server/module-api/src/main/java/project/BaekjoonStatus/api/stat/service/cache/CacheCallback.java
@@ -1,0 +1,5 @@
+package project.BaekjoonStatus.api.stat.service.cache;
+
+public interface CacheCallback<T> {
+    T execute();
+}
