+++ server/module-api/src/main/java/project/BaekjoonStatus/api/stat/service/cache/CacheNode.java
@@ -1,0 +1,16 @@
+package project.BaekjoonStatus.api.stat.service.cache;
+
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+@Data
+public class CacheNode<T> {
+    LocalDateTime expireAt;
+    T data;
+
+    public CacheNode(LocalDateTime expireAt, T data) {
+        this.expireAt = expireAt;
+        this.data = data;
+    }
+}
